+++ src/Server/Process/Background.php
@@ -6,14 +6,14 @@ namespace Innmind\Server\Control\Server\Process;
 use Innmind\Server\Control\Server\Process;
 use Innmind\Immutable\{
     Sequence,
-    Str,
     Maybe,
     Either,
 };
 
 final class Background implements Process
 {
-    private Output $output;
+    /** @var Sequence<Output\Chunk> */
+    private Sequence $output;
 
     public function __construct(Started $process)
     {
@@ -21,9 +21,7 @@ final class Background implements Process
         // process will be killed
         // this also allows to send any input to the stream
         $process->output()->memoize();
-        /** @var Sequence<array{0: Str, 1: Output\Type}> */
-        $output = Sequence::of();
-        $this->output = new Output\Output($output);
+        $this->output = Sequence::of();
 
         // the pid returned by `$process->pid()` is the one for the "foreground"
         // process that starts the background one, the real pid (the one we
@@ -42,7 +40,7 @@ final class Background implements Process
     }
 
     #[\Override]
-    public function output(): Output
+    public function output(): Sequence
     {
         return $this->output;
     }