+++ tests/Server/Process/ForegroundTest.php
@@ -7,7 +7,6 @@ use Innmind\Server\Control\Server\{
     Process\Foreground,
     Process\Unix,
     Process,
-    Process\Output,
     Process\Output\Type,
     Process\Failed,
     Process\Success,
@@ -19,7 +18,7 @@ use Innmind\TimeContinuum\Earth\{
 };
 use Innmind\TimeWarp\Halt\Usleep;
 use Innmind\Stream\Streams;
-use Innmind\Immutable\Str;
+use Innmind\Immutable\Monoid\Concat;
 use PHPUnit\Framework\TestCase;
 
 class ForegroundTest extends TestCase
@@ -72,21 +71,27 @@ class ForegroundTest extends TestCase
         );
         $process = new Foreground($slow());
 
-        $this->assertInstanceOf(Output::class, $process->output());
         $start = \time();
         $count = 0;
         $process
             ->output()
-            ->foreach(function(Str $data, Type $type) use ($start, &$count) {
-                $this->assertSame($count."\n", $data->toString());
+            ->foreach(function($chunk) use ($start, &$count) {
+                $this->assertSame($count."\n", $chunk->data()->toString());
                 $this->assertEquals(
-                    (int) $data->toString() % 2 === 0 ? Type::output : Type::error,
-                    $type,
+                    (int) $chunk->data()->toString() % 2 === 0 ? Type::output : Type::error,
+                    $chunk->type(),
                 );
                 $this->assertTrue((\time() - $start) >= (1 + $count));
                 ++$count;
             });
-        $this->assertSame("0\n1\n2\n3\n4\n5\n", $process->output()->toString());
+        $this->assertSame(
+            "0\n1\n2\n3\n4\n5\n",
+            $process
+                ->output()
+                ->map(static fn($chunk) => $chunk->data())
+                ->fold(new Concat)
+                ->toString(),
+        );
         $this->assertSame(6, $count);
     }
 
@@ -187,7 +192,7 @@ class ForegroundTest extends TestCase
                 ->withEnvironment('PATH', $_SERVER['PATH']),
         );
         $process = new Foreground($slow());
-        $process->output()->toString();
+        $process->output()->memoize();
 
         $this->assertInstanceOf(
             Success::class,
@@ -220,6 +225,13 @@ class ForegroundTest extends TestCase
                     static fn() => null,
                 ),
         );
-        $this->assertSame("0\n1\n2\n3\n4\n5\n", $process->output()->toString());
+        $this->assertSame(
+            "0\n1\n2\n3\n4\n5\n",
+            $process
+                ->output()
+                ->map(static fn($chunk) => $chunk->data())
+                ->fold(new Concat)
+                ->toString(),
+        );
     }
 }